requirements.txt
pyppeteer
pyzbar
qrcode
Pillow
